# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Mazayada API Documentation'
  description: 'Mazayada mobile API (v1). Token-authenticated JSON API for the Flutter client covering auctions and the citizen dashboard. All responses use a unified envelope ({data, message, meta} on success; {message, errors} on error). Money is returned in dinars; send Accept-Language (ar|fr|en) to localize responses.'
  version: 1.0.0
servers:
  -
    url: 'https://mazayada.findosystem.com'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'Obtain an <b>access token</b> from <code>POST /api/v1/auth/login</code> or <code>POST /api/v1/auth/verify-otp</code> and send it as <code>Authorization: Bearer &lt;access_token&gt;</code>. Access tokens are short-lived — use the <b>refresh token</b> at <code>POST /api/v1/auth/refresh</code> (send the refresh token as the bearer) to obtain a new pair.'
security:
  -
    default: []
paths:
...
  '/api/v1/auctions/{auction_id}/bids':
    get:
      summary: 'Latest bids'
      operationId: latestBids
      description: "The most recent valid bids (alias only) plus the current price. Use this to\nhydrate the bid history, or poll it as a fallback when the realtime socket\nis unavailable."
      parameters:
        -
          in: query
          name: limit
          description: 'Number of bids to return (1-50, default 10).'
          example: 10
          required: false
          schema:
            type: integer
            description: 'Number of bids to return (1-50, default 10).'
            example: 10
        -
          in: header
          name: Accept-Language
          description: ''
          example: ar
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    -
                      amount:
                        amount: 6100000
                        formatted: '6 100 000 دج'
                      bidder_alias: Bold_Falcon_197
                      bid_time: '2026-06-14T10:47:07+01:00'
                    -
                      amount:
                        amount: 6000000
                        formatted: '6 000 000 دج'
                      bidder_alias: Swift_Hawk_242
                      bid_time: '2026-06-14T10:46:30+01:00'
                    -
                      amount:
                        amount: 5500000
                        formatted: '5 500 000 دج'
                      bidder_alias: Swift_Hawk_242
                      bid_time: '2026-06-14T09:41:59+01:00'
                    -
                      amount:
                        amount: 5400000
                        formatted: '5 400 000 دج'
                      bidder_alias: Swift_Hawk_242
                      bid_time: '2026-06-14T09:41:37+01:00'
                    -
                      amount:
                        amount: 5300000
                        formatted: '5 300 000 دج'
                      bidder_alias: Bold_Falcon_197
                      bid_time: '2026-06-14T09:41:22+01:00'
                    -
                      amount:
                        amount: 5200000
                        formatted: '5 200 000 دج'
                      bidder_alias: Swift_Hawk_242
                      bid_time: '2026-06-14T09:41:12+01:00'
                    -
                      amount:
                        amount: 5000000
                        formatted: '5 000 000 دج'
                      bidder_alias: Swift_Hawk_242
                      bid_time: '2026-06-14T09:41:04+01:00'
                    -
                      amount:
                        amount: 4600000
                        formatted: '4 600 000 دج'
                      bidder_alias: Bold_Falcon_197
                      bid_time: '2026-06-14T09:40:56+01:00'
                    -
                      amount:
                        amount: 3500000
                        formatted: '3 500 000 دج'
                      bidder_alias: Bold_Falcon_197
                      bid_time: '2026-06-14T09:40:45+01:00'
                    -
                      amount:
                        amount: 3400000
                        formatted: '3 400 000 دج'
                      bidder_alias: Swift_Hawk_242
                      bid_time: '2026-06-14T09:40:32+01:00'
                  message: null
                  meta:
                    current_price: 6100000
                    bid_count: 15
                    status: CLOSED
                properties:
                  data:
                    type: array
                    example:
                      -
                        amount:
                          amount: 6100000
                          formatted: '6 100 000 دج'
                        bidder_alias: Bold_Falcon_197
                        bid_time: '2026-06-14T10:47:07+01:00'
                      -
                        amount:
                          amount: 6000000
                          formatted: '6 000 000 دج'
                        bidder_alias: Swift_Hawk_242
                        bid_time: '2026-06-14T10:46:30+01:00'
                      -
                        amount:
                          amount: 5500000
                          formatted: '5 500 000 دج'
                        bidder_alias: Swift_Hawk_242
                        bid_time: '2026-06-14T09:41:59+01:00'
                      -
                        amount:
                          amount: 5400000
                          formatted: '5 400 000 دج'
                        bidder_alias: Swift_Hawk_242
                        bid_time: '2026-06-14T09:41:37+01:00'
                      -
                        amount:
                          amount: 5300000
                          formatted: '5 300 000 دج'
                        bidder_alias: Bold_Falcon_197
                        bid_time: '2026-06-14T09:41:22+01:00'
                      -
                        amount:
                          amount: 5200000
                          formatted: '5 200 000 دج'
                        bidder_alias: Swift_Hawk_242
                        bid_time: '2026-06-14T09:41:12+01:00'
                      -
                        amount:
                          amount: 5000000
                          formatted: '5 000 000 دج'
                        bidder_alias: Swift_Hawk_242
                        bid_time: '2026-06-14T09:41:04+01:00'
                      -
                        amount:
                          amount: 4600000
                          formatted: '4 600 000 دج'
                        bidder_alias: Bold_Falcon_197
                        bid_time: '2026-06-14T09:40:56+01:00'
                      -
                        amount:
                          amount: 3500000
                          formatted: '3 500 000 دج'
                        bidder_alias: Bold_Falcon_197
                        bid_time: '2026-06-14T09:40:45+01:00'
                      -
                        amount:
                          amount: 3400000
                          formatted: '3 400 000 دج'
                        bidder_alias: Swift_Hawk_242
                        bid_time: '2026-06-14T09:40:32+01:00'
                    items:
                      type: object
                      properties:
                        amount:
                          type: object
                          properties:
                            amount:
                              type: integer
                              example: 6100000
                            formatted:
                              type: string
                              example: '6 100 000 دج'
                        bidder_alias:
                          type: string
                          example: Bold_Falcon_197
                        bid_time:
                          type: string
                          example: '2026-06-14T10:47:07+01:00'
                  message:
                    type: string
                    example: null
                    nullable: true
                  meta:
                    type: object
                    properties:
                      current_price:
                        type: integer
                        example: 6100000
                      bid_count:
                        type: integer
                        example: 15
                      status:
                        type: string
                        example: CLOSED
      tags:
        - Auctions
      security: []
    parameters:
      -
        in: path
        name: auction_id
        description: 'The ID of the auction.'
        example: a2050443-68af-41f5-8080-20d4b878c6ca
        required: true
        schema:
          type: string
  '/api/v1/auctions/{auction_id}/price':
    get:
      summary: 'Current price'
      operationId: currentPrice
      description: 'A cheap, pollable snapshot of the live price and clock for an auction.'
      parameters:
        -
          in: header
          name: Accept-Language
          description: ''
          example: ar
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    current_price: 6100000
                    current_price_formatted: '6 100 000 دج'
                    bid_count: 15
                    status: CLOSED
                    end_time: '2026-06-14T12:00:00+01:00'
                    is_biddable: false
                    has_ended: false
                  message: null
                  meta: {  }
                properties:
                  data:
                    type: object
                    properties:
                      current_price:
                        type: integer
                        example: 6100000
                      current_price_formatted:
                        type: string
                        example: '6 100 000 دج'
                      bid_count:
                        type: integer
                        example: 15
                      status:
                        type: string
                        example: CLOSED
                      end_time:
                        type: string
                        example: '2026-06-14T12:00:00+01:00'
                      is_biddable:
                        type: boolean
                        example: false
                      has_ended:
                        type: boolean
                        example: false
                  message:
                    type: string
                    example: null
                    nullable: true
                  meta:
                    type: object
                    properties: {  }
      tags:
        - Auctions
      security: []
    parameters:
      -
        in: path
        name: auction_id
        description: 'The ID of the auction.'
        example: a2050443-68af-41f5-8080-20d4b878c6ca
        required: true
        schema:
          type: string
...